Restructure frame module
This makes frame a more self contained module. Also, this adds
util/h2-codec, which contains a crate that has a goal of only exporting
the `Codec` types.
