+++ src/error.rs
@@ -1,5 +1,7 @@
-use http;
+use {frame, http};
 use std::{error, fmt, io};
+
+pub use frame::Reason;
 
 /// The error type for HTTP/2 operations
 ///
@@ -42,26 +44,6 @@ impl From<Reason> for StreamError {
 }
 
 #[derive(Debug, PartialEq, Eq, Clone, Copy)]
-pub enum Reason {
-    NoError,
-    ProtocolError,
-    InternalError,
-    FlowControlError,
-    SettingsTimeout,
-    StreamClosed,
-    FrameSizeError,
-    RefusedStream,
-    Cancel,
-    CompressionError,
-    ConnectError,
-    EnhanceYourCalm,
-    InadequateSecurity,
-    Http11Required,
-    Other(u32),
-    // TODO: reserve additional variants
-}
-
-#[derive(Debug, PartialEq, Eq, Clone, Copy)]
 pub enum User {
     /// The specified stream ID is invalid.
     ///
@@ -92,31 +74,6 @@ pub enum User {
     Rejected,
 
     // TODO: reserve additional variants
-}
-
-macro_rules! reason_desc {
-    ($reason:expr) => (reason_desc!($reason, ""));
-    ($reason:expr, $prefix:expr) => ({
-        use self::Reason::*;
-
-        match $reason {
-            NoError => concat!($prefix, "not a result of an error"),
-            ProtocolError => concat!($prefix, "unspecific protocol error detected"),
-            InternalError => concat!($prefix, "unexpected internal error encountered"),
-            FlowControlError => concat!($prefix, "flow-control protocol violated"),
-            SettingsTimeout => concat!($prefix, "settings ACK not received in timely manner"),
-            StreamClosed => concat!($prefix, "received frame when stream half-closed"),
-            FrameSizeError => concat!($prefix, "frame sent with invalid size"),
-            RefusedStream => concat!($prefix, "refused stream before processing any application logic"),
-            Cancel => concat!($prefix, "stream no longer needed"),
-            CompressionError => concat!($prefix, "unable to maintain the header compression context"),
-            ConnectError => concat!($prefix, "connection established in response to a CONNECT request was reset or abnormally closed"),
-            EnhanceYourCalm => concat!($prefix, "detected excessive load generating behavior"),
-            InadequateSecurity => concat!($prefix, "security properties do not meet minimum requirements"),
-            Http11Required => concat!($prefix, "endpoint requires HTTP/1.1"),
-            Other(_) => concat!($prefix, "other reason (ain't no tellin')"),
-        }
-    });
 }
 
 macro_rules! user_desc {
@@ -170,6 +127,15 @@ impl From<http::Error> for ConnectionError {
     }
 }
 
+impl From<frame::Error> for ConnectionError {
+    fn from(src: frame::Error) -> ConnectionError {
+        match src {
+            // TODO: implement
+            _ => ConnectionError::Proto(Reason::ProtocolError),
+        }
+    }
+}
+
 impl fmt::Display for ConnectionError {
     fn fmt(&self, fmt: &mut fmt::Formatter) -> fmt::Result {
         use self::ConnectionError::*;
@@ -188,71 +154,9 @@ impl error::Error for ConnectionError {
 
         match *self {
             Io(ref e) => error::Error::description(e),
-            Proto(reason) => reason_desc!(reason, "protocol error: "),
+            Proto(ref reason) => reason.description(),
             User(user) => user_desc!(user, "user error: "),
         }
-    }
-}
-
-// ===== impl Reason =====
-
-impl Reason {
-    pub fn description(&self) -> &str {
-        reason_desc!(*self)
-    }
-}
-
-impl From<u32> for Reason {
-    fn from(src: u32) -> Reason {
-        use self::Reason::*;
-
-        match src {
-            0x0 => NoError,
-            0x1 => ProtocolError,
-            0x2 => InternalError,
-            0x3 => FlowControlError,
-            0x4 => SettingsTimeout,
-            0x5 => StreamClosed,
-            0x6 => FrameSizeError,
-            0x7 => RefusedStream,
-            0x8 => Cancel,
-            0x9 => CompressionError,
-            0xa => ConnectError,
-            0xb => EnhanceYourCalm,
-            0xc => InadequateSecurity,
-            0xd => Http11Required,
-            _ => Other(src),
-        }
-    }
-}
-
-impl From<Reason> for u32 {
-    fn from(src: Reason) -> u32 {
-        use self::Reason::*;
-
-        match src {
-            NoError => 0x0,
-            ProtocolError => 0x1,
-            InternalError => 0x2,
-            FlowControlError => 0x3,
-            SettingsTimeout => 0x4,
-            StreamClosed => 0x5,
-            FrameSizeError => 0x6,
-            RefusedStream => 0x7,
-            Cancel => 0x8,
-            CompressionError => 0x9,
-            ConnectError => 0xa,
-            EnhanceYourCalm => 0xb,
-            InadequateSecurity => 0xc,
-            Http11Required => 0xd,
-            Other(v) => v,
-        }
-    }
-}
-
-impl fmt::Display for Reason {
-    fn fmt(&self, fmt: &mut fmt::Formatter) -> fmt::Result {
-        write!(fmt, "{}", self.description())
     }
 }
