+++ util/h2-codec/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "h2-codec"
+version = "0.1.0"
+authors = ["Carl Lerche <me@carllerche.com>"]
+
+[dependencies]
+http = { git = "https://github.com/carllerche/http" }
+log = "0.3.8"
+fnv = "1.0.5"
+bytes = "0.4"
+string = { git = "https://github.com/carllerche/string" }
+byteorder = "1.0"
+
+# futures = "0.1"
+# tokio-io = "0.1.3"
+# tokio-timer = "0.1"
+# http = { git = "https://github.com/carllerche/http", branch = "lower-case-header-name-parsing" }
+# slab = "0.4.0"
