+++ util/h2-codec/src/lib.rs
@@ -1,0 +1,14 @@
+extern crate http;
+extern crate fnv;
+extern crate bytes;
+extern crate string;
+extern crate byteorder;
+
+#[macro_use]
+extern crate log;
+
+#[path = "../../../src/hpack/mod.rs"]
+mod hpack;
+
+#[path = "../../../src/frame/mod.rs"]
+mod frame;
